+++ Server/src/main/java/Server/core/game/node/entity/npc/city/alkharid/ZekeDialogue.java
@@ -1,5 +1,7 @@
-package core.game.content.dialogue;
+package core.game.node.entity.npc.city.alkharid;
 
+import core.game.content.dialogue.DialoguePlugin;
+import core.game.content.dialogue.FacialExpression;
 import core.game.node.entity.npc.NPC;
 import core.plugin.Initializable;
 import core.game.node.entity.player.Player;
